backend/requirements.txt
flask
flask-cors
python-dotenv
mysql-connector-python
PyJWT>=2.0.0
